[desktop] Redesign Create new Vault flow (rename as well)
