+++ src/containers/Modal/CreateOrEditVaultModalContentV2/CreateOrEditVaultModalContentV2.styles.ts
@@ -6,14 +6,5 @@ export const createStyles = () => ({
     flexDirection: 'column' as const,
     gap: `${rawTokens.spacing16}px`,
     width: '100%'
-  },
-  passwordFields: {
-    display: 'flex' as const,
-    flexDirection: 'column' as const,
-    gap: `${rawTokens.spacing12}px`,
-    width: '100%'
-  },
-  toggleWrap: {
-    width: '100%'
   }
 })
