+++ src/containers/Modal/CreateOrEditVaultModalContentV2/CreateOrEditVaultModalContentV2.tsx
@@ -1,0 +1,155 @@
+import React, { useEffect, useState } from 'react'
+
+import { useForm } from '@tetherto/pear-apps-lib-ui-react-hooks'
+import { Validator } from '@tetherto/pear-apps-utils-validator'
+import { Button, Dialog, Form, InputField } from '@tetherto/pearpass-lib-ui-kit'
+import { useCreateVault, useVault, type Vault } from '@tetherto/pearpass-lib-vault'
+
+import { createStyles } from './CreateOrEditVaultModalContentV2.styles'
+import { useGlobalLoading } from '../../../context/LoadingContext'
+import { useRouter } from '../../../context/RouterContext'
+import { useTranslation } from '../../../hooks/useTranslation'
+import { getDeviceName } from '../../../utils/getDeviceName'
+import { logger } from '../../../utils/logger'
+
+export type CreateOrEditVaultModalContentV2Props = {
+  onClose: () => void
+  onSuccess?: () => void
+  vault?: Vault
+}
+
+export const CreateOrEditVaultModalContentV2 = ({
+  onClose,
+  onSuccess,
+  vault
+}: CreateOrEditVaultModalContentV2Props) => {
+  const isEditMode = !!vault
+  const { t } = useTranslation()
+  const { navigate } = useRouter()
+  const styles = createStyles()
+  const [isLoading, setIsLoading] = useState(false)
+
+  useGlobalLoading({ isLoading })
+
+  const schema = Validator.object({
+    name: Validator.string().required(t('Name is required'))
+  })
+
+  const { addDevice, updateUnprotectedVault, refetch: refetchVault } = useVault()
+  const { createVault } = useCreateVault()
+
+  const { register, handleSubmit, values } = useForm({
+    initialValues: {
+      name: vault?.name ?? ''
+    },
+    validate: (formValues: { name: string }) => schema.validate(formValues)
+  })
+
+  useEffect(() => {
+    if (isEditMode) {
+      void refetchVault()
+    }
+  }, [isEditMode, refetchVault])
+
+  const nameField = register('name')
+
+  const nameOk = values.name.trim().length > 0
+  const isSaveDisabled = !nameOk || isLoading
+
+  const submit = async (formValues: { name: string }) => {
+    if (isLoading) {
+      return
+    }
+
+    if (isEditMode && vault) {
+      try {
+        setIsLoading(true)
+        await updateUnprotectedVault(vault.id, { name: formValues.name })
+        setIsLoading(false)
+        onSuccess?.()
+      } catch (error) {
+        setIsLoading(false)
+        logger.error(
+          'CreateOrEditVaultModalContentV2',
+          'Error renaming vault:',
+          error
+        )
+      }
+      return
+    }
+
+    try {
+      setIsLoading(true)
+
+      await createVault({
+        name: formValues.name,
+        password: ''
+      })
+
+      await addDevice(getDeviceName())
+
+      onSuccess?.()
+      navigate('vault', { recordType: 'all' })
+
+      setIsLoading(false)
+    } catch (error) {
+      setIsLoading(false)
+      logger.error(
+        'CreateOrEditVaultModalContentV2',
+        'Error creating vault:',
+        error
+      )
+    }
+  }
+
+  const nameError = nameField.error
+
+  return (
+    <Dialog
+      title={isEditMode ? t('Rename Vault') : t('Create New Vault')}
+      onClose={onClose}
+      testID="createvault-dialog-v2"
+      closeButtonTestID="createvault-close-v2"
+      footer={
+        <>
+          <Button
+            variant="secondary"
+            size="small"
+            type="button"
+            onClick={onClose}
+            data-testid="createvault-discard-v2"
+          >
+            {t('Discard')}
+          </Button>
+          <Button
+            variant="primary"
+            size="small"
+            type="button"
+            disabled={isSaveDisabled}
+            isLoading={isLoading}
+            onClick={() => handleSubmit(submit)()}
+            data-testid="createvault-save-v2"
+          >
+            {t('Save')}
+          </Button>
+        </>
+      }
+    >
+      <Form
+        onSubmit={handleSubmit(submit)}
+        style={styles.form as React.ComponentProps<typeof Form>['style']}
+        testID="createvault-form-v2"
+      >
+        <InputField
+          label={t('Vault Name')}
+          placeholderText={t('Enter Name')}
+          value={nameField.value}
+          onChangeText={(v) => nameField.onChange(v)}
+          variant={nameError ? 'error' : 'default'}
+          errorMessage={nameError || undefined}
+          testID="createvault-name-v2"
+        />
+      </Form>
+    </Dialog>
+  )
+}
