+++ src/pages/WelcomePage/CardNewVaultCredentials/index.js
@@ -3,8 +3,8 @@ import { useEffect } from 'react'
 import { html } from 'htm/react'
 
 import { NAVIGATION_ROUTES } from '../../../constants/navigation'
+import { CreateOrEditVaultModalContentV2 } from '../../../containers/Modal/CreateOrEditVaultModalContentV2/CreateOrEditVaultModalContentV2'
 import { CreateVaultModalContent } from '../../../containers/Modal/CreateVaultModalContent'
-import { CreateVaultModalContentV2 } from '../../../containers/Modal/CreateVaultModalContentV2/CreateVaultModalContentV2'
 import { useModal } from '../../../context/ModalContext'
 import { useRouter } from '../../../context/RouterContext'
 import { isV2 } from '../../../utils/designVersion'
@@ -20,7 +20,7 @@ export const CardNewVaultCredentials = () => {
     }
 
     const CreateContent = isV2()
-      ? CreateVaultModalContentV2
+      ? CreateOrEditVaultModalContentV2
       : CreateVaultModalContent
 
     setModal(
